Check the examples in Problem 9
I think the example 2 and example 7 should return 'YES'. Both points are within the circle `K( (1,1), 3)`, but outside the rectangle `R(top=1, left=-1, width=6, height=2)`. The rectangle is situated between X = -1 and 5 and Y = -1 and 1. The point in example 2 is at Y = 2 (outside the rectangle). The point in example 7 is at Y = 1.5 (again out of the rect).